--- src/components/icons/BabyBottle.tsx
@@ -1,13 +1,0 @@
-import { IconProps } from '@tsc/theaterDetail/Icon';
-
-export const BabyBottle = ({ width = '50px', height = '50px', color = 'white' }: IconProps) => {
-  return (
-    <svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg " width={width} height={height} fill={color}>
-      <title />
-      <g id="Child_wellness">
-        <path d="M460.6412,85.2665a24.9179,24.9179,0,0,0-35.2382-35.24l-23.69,23.6895c-5.9551,5.9551-14.9453,8.15-22.7305,4.9395a98.149,98.149,0,0,0-76.31.6718,25.7984,25.7984,0,0,0-36.4834,0l-27.81,27.8106L403.53,272.29l27.8106-27.81a25.8,25.8,0,0,0,0-36.4844,98.15,98.15,0,0,0,.6728-76.3106c-3.2119-7.7871-1.0166-16.7734,4.94-22.73Z" />
-        <path d="M201.2037,151.077,165.4908,186.79l30.68,30.6817a11.999,11.999,0,1,1-16.97,16.9687l-30.6807-30.68-43.0234,43.0235,30.6816,30.68a12,12,0,0,1-16.9707,16.9707l-30.6816-30.68L54.18,298.1005a34.5457,34.5457,0,0,0-.002,48.8554L166.3776,459.1532a34.5446,34.5446,0,0,0,48.8545,0L362.2555,312.13a130.01,130.01,0,0,0,19.6279-25.41L226.6139,131.45A129.9409,129.9409,0,0,0,201.2037,151.077Z" />
-      </g>
-    </svg>
-  );
-};
