--- src/components/icons/Cafe.tsx
@@ -1,30 +1,0 @@
-import { IconProps } from "@tsc/theaterDetail/Icon";
-
-  export const CafeIcon = ({
-    width = "50px",
-    height = "50px",
-    color = "white",
-  }: IconProps) => {
-    return (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        enableBackground="new 0 0 48 48"
-        version="1.1"
-        viewBox="0 0 48 48"
-        width={width}
-        height={height}
-        fill={color}
-      >
-        <g id="Layer_3">
-          <g>
-            <path d="M0.022,40.362c0.418,1.307,0.935,2.569,1.549,3.775h44.856c0.615-1.206,1.133-2.469,1.551-3.775H0.022z" />
-            <path d="M13.931,37.845h12.587c2.963,0,5.434-2.057,6.1-4.817c1.666-0.798,3.953-2.113,6.162-4.119    c0.041-0.04,0.088-0.073,0.127-0.113c0.006-0.006,0.014-0.011,0.02-0.018c0-0.002,0-0.004,0-0.006    c0.885-0.907,1.436-2.145,1.436-3.513c0-2.78-2.254-5.034-5.035-5.034h-2.516V17.76H7.638v13.792    C7.638,35.027,10.456,37.845,13.931,37.845z M32.811,22.741h2.516c1.393,0,2.518,1.128,2.518,2.518    c0,0.366-0.082,0.71-0.225,1.022c-0.043,0.091-0.182,0.336-0.479,0.688c-0.059,0.063-0.113,0.134-0.18,0.19l-0.006,0.023    c-0.656,0.708-1.893,1.735-4.145,2.806V22.741z" />
-            <path d="M14.356,14.66c0.093,0.186,0.199,0.361,0.31,0.529h1.171c-0.007-0.014-0.017-0.024-0.024-0.038    c-0.878-1.727-0.731-3.702,0.208-5.243c0.127-0.162,0.244-0.331,0.35-0.505c0.02-0.026,0.037-0.052,0.058-0.075    c0.171-0.284,0.301-0.584,0.401-0.873c0.406-1.107,0.454-2.354,0.047-3.56c-0.125-0.37-0.291-0.713-0.483-1.033h-1.043    c0.05,0.119,0.098,0.239,0.141,0.363c0.554,1.646,0.263,3.369-0.632,4.709h0.006c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.599,1.296c-0.104,0.304-0.186,0.633-0.222,0.966c0.006-0.003,0.011-0.007,0.017-0.008    C13.689,12.611,13.855,13.677,14.356,14.66z" />
-            <path d="M19.197,14.66c0.094,0.186,0.199,0.361,0.31,0.529h1.171c-0.007-0.014-0.017-0.024-0.023-0.038    c-0.878-1.727-0.732-3.702,0.208-5.243c0.126-0.162,0.244-0.331,0.35-0.505c0.02-0.026,0.037-0.052,0.057-0.075    c0.171-0.284,0.301-0.584,0.401-0.873c0.406-1.107,0.454-2.354,0.047-3.56c-0.125-0.37-0.292-0.713-0.483-1.033h-1.043    c0.05,0.119,0.098,0.239,0.141,0.363c0.554,1.646,0.263,3.369-0.632,4.709h0.007c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.6,1.296c-0.103,0.304-0.186,0.633-0.221,0.966c0.005-0.003,0.011-0.007,0.016-0.008    C18.529,12.611,18.697,13.677,19.197,14.66z" />
-            <path d="M24.039,14.66c0.093,0.186,0.198,0.361,0.31,0.529h1.171c-0.006-0.014-0.016-0.024-0.023-0.038    c-0.878-1.727-0.731-3.702,0.209-5.243c0.125-0.162,0.244-0.331,0.348-0.505c0.02-0.026,0.037-0.052,0.057-0.075    c0.172-0.284,0.303-0.584,0.404-0.873c0.406-1.107,0.451-2.354,0.045-3.56c-0.123-0.37-0.289-0.713-0.48-1.033h-1.045    c0.051,0.119,0.1,0.239,0.141,0.363c0.555,1.646,0.266,3.369-0.631,4.709h0.007c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.6,1.296c-0.103,0.304-0.186,0.633-0.221,0.966c0.006-0.003,0.011-0.007,0.016-0.008    C23.37,12.611,23.538,13.677,24.039,14.66z" />
-          </g>
-        </g>
-      </svg>
-    );
-  };
-
