+++ src/components/icons/DetailIcons.tsx
@@ -1,0 +1,159 @@
+import { IconProps } from '@tsc/theaterDetail/Icon';
+
+export const BabyBottle = ({ width = '50px', height = '50px', color = 'white' }: IconProps) => {
+    return (
+      <svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg " width={width} height={height} fill={color}>
+        <title />
+        <g id="Child_wellness">
+          <path d="M460.6412,85.2665a24.9179,24.9179,0,0,0-35.2382-35.24l-23.69,23.6895c-5.9551,5.9551-14.9453,8.15-22.7305,4.9395a98.149,98.149,0,0,0-76.31.6718,25.7984,25.7984,0,0,0-36.4834,0l-27.81,27.8106L403.53,272.29l27.8106-27.81a25.8,25.8,0,0,0,0-36.4844,98.15,98.15,0,0,0,.6728-76.3106c-3.2119-7.7871-1.0166-16.7734,4.94-22.73Z" />
+          <path d="M201.2037,151.077,165.4908,186.79l30.68,30.6817a11.999,11.999,0,1,1-16.97,16.9687l-30.6807-30.68-43.0234,43.0235,30.6816,30.68a12,12,0,0,1-16.9707,16.9707l-30.6816-30.68L54.18,298.1005a34.5457,34.5457,0,0,0-.002,48.8554L166.3776,459.1532a34.5446,34.5446,0,0,0,48.8545,0L362.2555,312.13a130.01,130.01,0,0,0,19.6279-25.41L226.6139,131.45A129.9409,129.9409,0,0,0,201.2037,151.077Z" />
+        </g>
+      </svg>
+    );
+  };
+
+
+  export const CafeIcon = ({
+    width = "50px",
+    height = "50px",
+    color = "white",
+  }: IconProps) => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        enableBackground="new 0 0 48 48"
+        version="1.1"
+        viewBox="0 0 48 48"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <g id="Layer_3">
+          <g>
+            <path d="M0.022,40.362c0.418,1.307,0.935,2.569,1.549,3.775h44.856c0.615-1.206,1.133-2.469,1.551-3.775H0.022z" />
+            <path d="M13.931,37.845h12.587c2.963,0,5.434-2.057,6.1-4.817c1.666-0.798,3.953-2.113,6.162-4.119    c0.041-0.04,0.088-0.073,0.127-0.113c0.006-0.006,0.014-0.011,0.02-0.018c0-0.002,0-0.004,0-0.006    c0.885-0.907,1.436-2.145,1.436-3.513c0-2.78-2.254-5.034-5.035-5.034h-2.516V17.76H7.638v13.792    C7.638,35.027,10.456,37.845,13.931,37.845z M32.811,22.741h2.516c1.393,0,2.518,1.128,2.518,2.518    c0,0.366-0.082,0.71-0.225,1.022c-0.043,0.091-0.182,0.336-0.479,0.688c-0.059,0.063-0.113,0.134-0.18,0.19l-0.006,0.023    c-0.656,0.708-1.893,1.735-4.145,2.806V22.741z" />
+            <path d="M14.356,14.66c0.093,0.186,0.199,0.361,0.31,0.529h1.171c-0.007-0.014-0.017-0.024-0.024-0.038    c-0.878-1.727-0.731-3.702,0.208-5.243c0.127-0.162,0.244-0.331,0.35-0.505c0.02-0.026,0.037-0.052,0.058-0.075    c0.171-0.284,0.301-0.584,0.401-0.873c0.406-1.107,0.454-2.354,0.047-3.56c-0.125-0.37-0.291-0.713-0.483-1.033h-1.043    c0.05,0.119,0.098,0.239,0.141,0.363c0.554,1.646,0.263,3.369-0.632,4.709h0.006c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.599,1.296c-0.104,0.304-0.186,0.633-0.222,0.966c0.006-0.003,0.011-0.007,0.017-0.008    C13.689,12.611,13.855,13.677,14.356,14.66z" />
+            <path d="M19.197,14.66c0.094,0.186,0.199,0.361,0.31,0.529h1.171c-0.007-0.014-0.017-0.024-0.023-0.038    c-0.878-1.727-0.732-3.702,0.208-5.243c0.126-0.162,0.244-0.331,0.35-0.505c0.02-0.026,0.037-0.052,0.057-0.075    c0.171-0.284,0.301-0.584,0.401-0.873c0.406-1.107,0.454-2.354,0.047-3.56c-0.125-0.37-0.292-0.713-0.483-1.033h-1.043    c0.05,0.119,0.098,0.239,0.141,0.363c0.554,1.646,0.263,3.369-0.632,4.709h0.007c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.6,1.296c-0.103,0.304-0.186,0.633-0.221,0.966c0.005-0.003,0.011-0.007,0.016-0.008    C18.529,12.611,18.697,13.677,19.197,14.66z" />
+            <path d="M24.039,14.66c0.093,0.186,0.198,0.361,0.31,0.529h1.171c-0.006-0.014-0.016-0.024-0.023-0.038    c-0.878-1.727-0.731-3.702,0.209-5.243c0.125-0.162,0.244-0.331,0.348-0.505c0.02-0.026,0.037-0.052,0.057-0.075    c0.172-0.284,0.303-0.584,0.404-0.873c0.406-1.107,0.451-2.354,0.045-3.56c-0.123-0.37-0.289-0.713-0.48-1.033h-1.045    c0.051,0.119,0.1,0.239,0.141,0.363c0.555,1.646,0.266,3.369-0.631,4.709h0.007c0,0-0.109,0.156-0.254,0.403    c-0.256,0.404-0.459,0.84-0.6,1.296c-0.103,0.304-0.186,0.633-0.221,0.966c0.006-0.003,0.011-0.007,0.016-0.008    C23.37,12.611,23.538,13.677,24.039,14.66z" />
+          </g>
+        </g>
+      </svg>
+    );
+  };
+
+
+  export const ElevatorIcon = ({ width = '50px', height = '50px', color = 'white' }: IconProps) => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        xmlnsXlink="http://www.w3.org/1999/xlink"
+        viewBox="0 0 1024 1024"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <path
+          d="M881.1,90.6H142.9c-28.1,0-51,22.9-51,51v740.8c0,28.1,22.9,51,51,51h738.1c28.1,0,51-22.9,51-51V141.6 C932.1,113.5,909.2,90.6,881.1,90.6z M354.3,223.3c36.8,0,66.7,29.9,66.7,66.7s-29.9,66.7-66.7,66.7c-36.8,0-66.7-29.9-66.7-66.7 S317.5,223.3,354.3,223.3z M461,641.1h-40v132.6c0,15.3-11.8,27.4-26.8,27.4h-79.1c-15.5,0-28.1-12.3-28.1-27.4V641.1h-39V488.3 c0-24.2,9.3-46.9,26.1-63.8c16.9-17,39.5-26.4,63.7-26.4h33.1c24.2,0,46.9,9.4,63.8,26.4c17,17,26.3,39.7,26.3,63.9V641.1z M680.8,721.5L556.6,597.3l28.3-28.3l95.9,95.9l95.9-95.9l28.3,28.3L680.8,721.5z M776.7,455l-95.9-95.9L584.9,455l-28.3-28.3 l124.2-124.2L805,426.7L776.7,455z"
+          id="XMLID_189_"
+        />
+      </svg>
+    );
+  };
+  
+
+  export const ParkingIcon = ({
+    width = '50px',
+    height = '50px',
+    color = 'white',
+  }: IconProps) => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <g>
+          <path d="M0 0h24v24H0z" fill="none" />
+          <path d="M11 14h1.5a3.5 3.5 0 0 0 0-7H9v10h2v-3zM4 3h16a1 1 0 0 1 1 1v16a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1zm7 6h1.5a1.5 1.5 0 0 1 0 3H11V9z" />
+        </g>
+      </svg>
+    );
+  };
+
+
+  export const RestaurantIcon = ({
+    width = '50px',
+    height = '50px',
+    color = 'white',
+  }: IconProps) => {
+    return (
+      <svg
+        id="Layer_1"
+        viewBox="0 0 30 30"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlnsXlink="http://www.w3.org/1999/xlink"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <path d="M23,19l-3.328-3.232C19.239,15.273,19,14.637,19,13.98V12c0-4.945,3.157-9.535,3.157-9.535L23,2.999V19z" />
+        <circle cx="23" cy="3" r="1" />
+        <path d="M24,3h-2l-1,10v13.5c0,0.828,0.672,1.5,1.5,1.5h0c0.828,0,1.5-0.672,1.5-1.5V3z" />
+        <path d="M13.087,2.445C13.037,2.186,12.811,2,12.548,2C12.245,2,12,2.245,12,2.548v5.807C12,8.711,11.711,9,11.355,9  c-0.329,0-0.605-0.247-0.641-0.574l-0.66-5.939C10.023,2.21,9.789,2,9.509,2H9.5H9.491C9.211,2,8.977,2.21,8.946,2.488l-0.66,5.939  C8.25,8.753,7.974,9,7.645,9C7.289,9,7,8.711,7,8.355V2.548C7,2.245,6.755,2,6.452,2C6.189,2,5.963,2.186,5.913,2.445  C5.671,3.713,5,7.362,5,9c0,4,3,5,3,5v12.5C8,27.328,8.672,28,9.5,28s1.5-0.672,1.5-1.5V14c0,0,3-1,3-5  C14,7.362,13.329,3.713,13.087,2.445z" />
+      </svg>
+    );
+  };
+
+  export const RestroomIcon = ({ width = '50px', height = '50px', color = 'white' }: IconProps) => {
+    return (
+      <svg
+        viewBox="0 0 20 20"
+        xmlns="http://www.w3.org/2000/svg"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <path d="M12 16H9l2-4.5V9c0-1.1.9-2 2-2h2a2 2 0 0 1 2 2v2.5l2 4.5h-3v4h-4v-4zm-5-3h2V9a2 2 0 0 0-2-2H3a2 2 0 0 0-2 2v4h2v7h4v-7zM5 6a3 3 0 1 1 0-6 3 3 0 0 1 0 6zm9 0a3 3 0 1 1 0-6 3 3 0 0 1 0 6z" />
+      </svg>
+    );
+  };
+
+  export const SearchIcon = ({ width = '28px', height = '28px', color = 'white' }: IconProps) => {
+    return (
+      <svg
+        enableBackground="new 0 0 32 32"
+        height={height}
+        width={width}
+        viewBox="0 0 32 32"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlnsXlink="http://www.w3.org/1999/xlink"
+        fill={color}
+      >
+        <g>
+          <path d="M13,2C6.935,2,2,6.935,2,13s4.935,11,11,11s11-4.935,11-11S19.065,2,13,2z M13,22c-4.962,0-9-4.037-9-9   c0-4.962,4.038-9,9-9c4.963,0,9,4.038,9,9C22,17.963,17.963,22,13,22z" />
+          <path d="M29.707,28.293l-6.001-6c-0.391-0.391-1.023-0.391-1.414,0s-0.391,1.023,0,1.414l6.001,6C28.488,29.902,28.744,30,29,30   s0.512-0.098,0.707-0.293C30.098,29.316,30.098,28.684,29.707,28.293z" />
+        </g>
+      </svg>
+    );
+  };
+
+
+  export const WheelchairIcon = ({
+    width = '50px',
+    height = '50px',
+    color = 'white',
+  }: IconProps) => {
+    return (
+      <svg
+        viewBox="0 0 512 512"
+        xmlns="http://www.w3.org/2000/svg"
+        width={width}
+        height={height}
+        fill={color}
+      >
+        <path d="M496.101 385.669l14.227 28.663c3.929 7.915.697 17.516-7.218 21.445l-65.465 32.886c-16.049 7.967-35.556 1.194-43.189-15.055L331.679 320H192c-15.925 0-29.426-11.71-31.679-27.475C126.433 55.308 128.38 70.044 128 64c0-36.358 30.318-65.635 67.052-63.929 33.271 1.545 60.048 28.905 60.925 62.201.868 32.933-23.152 60.423-54.608 65.039l4.67 32.69H336c8.837 0 16 7.163 16 16v32c0 8.837-7.163 16-16 16H215.182l4.572 32H352a32 32 0 0 1 28.962 18.392L438.477 396.8l36.178-18.349c7.915-3.929 17.517-.697 21.446 7.218zM311.358 352h-24.506c-7.788 54.204-54.528 96-110.852 96-61.757 0-112-50.243-112-112 0-41.505 22.694-77.809 56.324-97.156-3.712-25.965-6.844-47.86-9.488-66.333C45.956 198.464 0 261.963 0 336c0 97.047 78.953 176 176 176 71.87 0 133.806-43.308 161.11-105.192L311.358 352z" />
+      </svg>
+    );
+  };
