--- src/components/icons/Elevator.tsx
@@ -1,21 +1,0 @@
-
-import { IconProps } from '@tsc/theaterDetail/Icon';
-
-export const ElevatorIcon = ({ width = '50px', height = '50px', color = 'white' }: IconProps) => {
-
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      xmlnsXlink="http://www.w3.org/1999/xlink"
-      viewBox="0 0 1024 1024"
-      width={width}
-      height={height}
-      fill={color}
-    >
-      <path
-        d="M881.1,90.6H142.9c-28.1,0-51,22.9-51,51v740.8c0,28.1,22.9,51,51,51h738.1c28.1,0,51-22.9,51-51V141.6 C932.1,113.5,909.2,90.6,881.1,90.6z M354.3,223.3c36.8,0,66.7,29.9,66.7,66.7s-29.9,66.7-66.7,66.7c-36.8,0-66.7-29.9-66.7-66.7 S317.5,223.3,354.3,223.3z M461,641.1h-40v132.6c0,15.3-11.8,27.4-26.8,27.4h-79.1c-15.5,0-28.1-12.3-28.1-27.4V641.1h-39V488.3 c0-24.2,9.3-46.9,26.1-63.8c16.9-17,39.5-26.4,63.7-26.4h33.1c24.2,0,46.9,9.4,63.8,26.4c17,17,26.3,39.7,26.3,63.9V641.1z M680.8,721.5L556.6,597.3l28.3-28.3l95.9,95.9l95.9-95.9l28.3,28.3L680.8,721.5z M776.7,455l-95.9-95.9L584.9,455l-28.3-28.3 l124.2-124.2L805,426.7L776.7,455z"
-        id="XMLID_189_"
-      />
-    </svg>
-  );
-};
