--- src/components/icons/Parking.tsx
@@ -1,23 +1,0 @@
-import { IconProps } from "@tsc/theaterDetail/Icon";
-  
-  export const ParkingIcon = ({
-    width = '50px',
-    height = '50px',
-    color = 'white',
-  }: IconProps) => {
-    return (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 24 24"
-        width={width}
-        height={height}
-        fill={color}
-      >
-        <g>
-          <path d="M0 0h24v24H0z" fill="none" />
-          <path d="M11 14h1.5a3.5 3.5 0 0 0 0-7H9v10h2v-3zM4 3h16a1 1 0 0 1 1 1v16a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1zm7 6h1.5a1.5 1.5 0 0 1 0 3H11V9z" />
-        </g>
-      </svg>
-    );
-  };
-
